+++ package.json
@@ -60,7 +60,8 @@
     "start": "rm -rf dist/  &&  parcel source/index.html --open",
     "pack-dist": "cross-env NODE_ENV=production  parcel build source/index.html --public-url .",
     "pack-sw": "rm -f dist/sw.js.map  &&  workbox generateSW",
-    "build": "rm -rf dist/  &&  npm run pack-dist  &&  npm run pack-sw"
+    "build": "rm -rf dist/  &&  npm run pack-dist  &&  npm run pack-sw",
+    "build-module": "parcel build source/index.ts --target node"
   },
   "husky": {
     "hooks": {
