--- source/demo.tsx
@@ -1,5 +1,0 @@
-import { documentReady, render, createCell } from 'web-cell';
-import { MapViz } from './page/MapViz';
-import './style.css';
-
-documentReady.then(() => render(<MapViz />));
