+++ source/index.html
@@ -19,7 +19,7 @@
     <script src="https://polyfill.io/v3/polyfill.min.js?flags=gated&features=Object.fromEntries%2CArray.prototype.flat"></script>
     <script src="https://cdn.jsdelivr.net/npm/@webcomponents/webcomponentsjs@2.4.1/custom-elements-es5-adapter.js"></script>
     <script src="https://cdn.jsdelivr.net/npm/@webcomponents/webcomponentsjs@2.4.1/webcomponents-bundle.js"></script>
-    <script src="demo.tsx" async></script>
+    <script src="index.tsx" async></script>
   </head>
   <body></body>
 </html>
