+++ source/index.tsx
@@ -1,14 +1,5 @@
+import { documentReady, render, createCell } from 'web-cell';
+import { MapViz } from './page/MapViz';
 import './style.css';
 
-export { HierarchicalVirusMap } from './components/HierarchicalVirusMap';
-export { InformationMap } from './components/InformationMap';
-
-export {
-  convertProvincesSeries,
-  convertCountrySeries
-} from './adapters/isaaclin';
-import patchData from './../data/isaaclin/patch.json';
-import rawData from './../data/isaaclin/history.json';
-import overviewData from './../data/isaaclin/overall.json';
-const VirusData = { patchData, rawData, overviewData };
-export { VirusData };
+documentReady.then(() => render(<MapViz />));
